requirements.txt
langchain>=0.0.200
langchain-openai>=0.0.2
langchain-community>=0.0.19
openai>=1.65.2
python-dotenv>=0.19.0
streamlit>=1.24.0
requests>=2.28.0
psycopg2-binary>=2.9.9
